API/pyproject.toml
[tool.poetry]
name = "gimvicurnik"
description = "An API for a school timetable, substitutions and menus at Gimnazija Vič"
readme = "README.md"
license = "GPL-3.0-or-later"
version = "0.0.0"

homepage = "https://gimvicurnik.filips.si/"
repository = "https://github.com/filips123/GimVicUrnik"
documentation = "https://github.com/filips123/GimVicUrnik/blob/main/README.md"

authors = [
    "filips123 <projects@filips.si>",
    "ChristofferNorgaard"
]

keywords = [
    "timetables",
    "substitutions",
    "menus",
    "gimvic"
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities"
]

[tool.poetry.scripts]
gimvicurnik = "gimvicurnik.__main__:main"

# TODO: Replace pdf2docx with PyMuPDF/pdfplumber/pdfminer
# TODO: Update to stable Flask once 2.2.0 is released
# TODO: Update to stable cattrs once 22.2.0 is released
[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0.0"
flask = { git = "https://github.com/pallets/flask.git", rev = "900e118" }
click = "^8.1.3"
schema = "^0.7.5"
attrs = "21.4.0"
cattrs = { git = "https://github.com/python-attrs/cattrs.git", rev = "56eda37" }
requests = "^2.28.1"
SQLAlchemy = "^1.4.39"
pdf2docx = "=0.4.6"
openpyxl = "^3.0.10"
icalendar = "^4.1.0"
beautifulsoup4 = { version = "^4.11.1", extras = ["lxml"] }

# Optional support for Sentry integration
sentry-sdk = { version = "^1.7.2", extras = ["flask", "sqlalchemy", "pure_eval"], optional = true }
setuptools = { version = "^63.2.0", optional = true }

# Optional support for different databases
PyMySQL = { version = "^1.0.2", optional = true }
mysqlclient = { version = "^2.1.1", optional = true }
psycopg2 = { version = "^2.9.3", optional = true }

[tool.poetry.dev-dependencies]
flakeheaven = "2.0.0"
black = "^22.6.0"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.7.1"
mypy = "^0.961"
typing-extensions = "^4.3.0"
sqlalchemy2-stubs = "^0.0.2-alpha.24"
types-setuptools = "^63.2.0"
types-pyyaml = "^6.0.10"
types-openpyxl = "^3.0.4"
types-requests = "^2.28.1"
types-beautifulsoup4 = "^4.11.3"

[tool.poetry.extras]
sentry = ["sentry-sdk", "setuptools"]
mysql-py = ["PyMySQL"]
mysql-c = ["mysqlclient"]
pgsql = ["psycopg2"]

[tool.flakeheaven]
max_line_length = 130
format = "grouped"

[tool.flakeheaven.plugins]
"flake8-black" = ["+*"]
"flake8-bugbear" = ["+*"]
"mccabe" = ["+*"]
"pycodestyle" = ["+*", "-E203", "-E501", "-W503"]
"pyflakes" = ["+*", "-F401"]

[tool.black]
line-length = 130

[tool.mypy]
python_version = 3.8
plugins = "sqlalchemy.ext.mypy.plugin"
show_column_numbers = true
show_error_codes = true
allow_redefinition = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"
